20190502- v3.08 Estimated Time: 0 Estimated Total Time: 32 : 25 Notes: Had to fix the not even starting up bug if files were not ther. Attempted to fix the new file 	deleting old file as well, could not find the issue.
=Bugs Squashed====================================================================================================
--20200411-0m |I dont think the version is saving correctly, it is current 0.05999999999 Instead of 0.06 On EKT
	Notes: Converted both _currentProjectVersion and _projectVersionIncrement to doubles and used
		Math.Round(x, 2)
--20200411-0m |If none of the recent loads are there, it crashes. Tries to load one of them and gets an exception.
	Notes: Fixed. Checked if File.Exists for each. Did not delete the files from recent loads so they can be
		seen.

+++ TODOList/DlgOptions.xaml.cs
@@ -12,11 +12,11 @@ namespace TODOList
 		public bool GlobalHotkeys { get; set; }
 		public bool AutoBackup { get; set; }
 		public TimeSpan BackupTime { get; set; }
-		public float CurrentProjectVersion { get; set; }
-		public float ProjectVersionIncrement { get; set; }
+		public double CurrentProjectVersion { get; set; }
+		public double ProjectVersionIncrement { get; set; }
 		public bool Result;
 		
-		public DlgOptions(bool autoSave, bool hotkeys, bool autoBackup, TimeSpan backupTime, float currentProjectVersion, float projectVersionIncrement)
+		public DlgOptions(bool autoSave, bool hotkeys, bool autoBackup, TimeSpan backupTime, double currentProjectVersion, double projectVersionIncrement)
 		{
 			InitializeComponent();
 			AutoSave = autoSave;
